ensure post is on table on request, create indexes in database

+++ bd/helpers.php
@@ -11,13 +11,17 @@ function getTableInfo($table) {
     global $config;
     return $config['tableinfo'][$table];
 }
+function exists_in_table($id, $table) {
+    return (bool) db_row("SELECT 1 FROM posts WHERE id = ? AND bd_table = ?", [$id, $table]);
+}
 function render_template($body = "<h1>test</h1>", $bar = "<a href='/'>/</a>") {
     ob_start();
     global $config;
-    $navbar = "<a href='/'>Home</a>";
+    $navbar = "<nav><a href='/'>Home</a>";
     foreach ($config['tables'] as $table) {
         $navbar = $navbar . " | <a href='/$table/'>/$table/</a>";
     }
+    $navbar = $navbar . "</nav>";
     $title = $config['title'];
     $desc = $config['description'];
     include $config['theme']; 
@@ -35,10 +39,12 @@ function deletePost($id, $delkey) {
 
     return false;
 }
-function deleteComment($id, $delkey) {
-    $row = db_row("SELECT delkey FROM replies WHERE id = ?", [$id]);
+function deleteComment($id, $delkey, $expectedtable) {
+    $row = db_row("SELECT delkey, post_id, table FROM replies WHERE id = ?", [$id]);
     if (!$row) return false;
-
+    if (!exists_in_table($row['post_id'], $expectedtable)) {
+        return false;
+    }
     if (password_verify($delkey, $row['delkey'])) {
         db_exec("DELETE FROM replies WHERE id = ?", [$id]);
         return true;
